[ENH] Refactor code to eliminate FileReader classes
Refactored the codebase, eliminating all instances of the FileReader and RawData classes. The data reading operation is updated to provide an UnstructuredData object directly. The RawDataOptions is removed and the reading function parameters are refactored. The unnecessary '__all__' cleanup is performed. A new enum for cells' special cases is introduced.