[ENH] Refactoring BoreholeSet initialization and improving coordinates combination
This commit refactors the BoreholeSet class initialization to simplify the handling of well IDs and coordinates. It removes the specificity of handling missing IDs, simplifying the handling to just raise an error or perform an intersect merge based on MergeOptions. Additionally, the process to add collar coordinates to each point of the survey trajectory during merging of DataFrame has been optimized. A slice parameter was also added to enable selection of portions of data during initialization.